+++ packages/modal-enhanced-react-native-web/src/index.js
@@ -1,7 +1,16 @@
 /**
  * MIT License
- * Copyright (c) 2017 React Native Community
- * https://github.com/react-native-community/react-native-modal
+ * Copyright (c) 2017 Ryan Florence
+ * https://github.com/reactjs/react-modal/blob/master/LICENSE
+ *
+ * Take WAI-ARIA for React Native Web
+ *
+ * Modified by Ray Andrew <raydreww@gmail.com>
+ * For Modal React Native Web
+ *
+ * MIT License
+ * Copyright (c) 2018 Ray Andrew
+ * https://github.com/rayandrews/react-native-web-modal
  */
 
 import React, { Component } from 'react';
@@ -21,7 +30,7 @@ import {
   registerAnimation,
   createAnimation,
 } from 'react-native-animatable';
-import { Modal } from 'modal-react-native-web';
+import Modal from 'modal-react-native-web';
 
 import * as ANIMATION_DEFINITIONS from './animations';
 
@@ -432,6 +441,7 @@ class ReactNativeModal extends Component {
       ) : (
         children
       );
+
     const containerView = (
       <View
         {...panHandlers}
