--- packages/modal-react-native-web/server.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/Portalv2/server');
