--- packages/modal-react-native-web/src/Portal/utils.js
@@ -1,5 +1,0 @@
-export const canUseDOM = !!(
-  typeof window !== 'undefined' &&
-  window.document &&
-  window.document.createElement
-);
