--- packages/modal-react-native-web/src/Portalv2/server/index.js
@@ -1,13 +1,0 @@
-import ReactDOMServer from 'react-dom/server';
-
-import { flushUniversalPortals } from '../client';
-
-export function getUniversalPortals() {
-  const portals = flushUniversalPortals();
-
-  if (!portals.length) {
-    return [];
-  }
-
-  return portals.map((children) => ReactDOMServer.renderToString(children));
-}
