--- packages/modal-react-native-web/src/Portalv3/index.js
@@ -1,2 +1,0 @@
-export { default as Modal } from './Modal';
-export { default as Portal } from './Portal';
